break up components, hooks and types
