+++ src/constants/problem.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_LANGUAGE = "cuda";
+export const DEFAULT_DATA_TYPE = "float32";
