+++ src/pages/problems/[slug].tsx
@@ -29,7 +29,7 @@ import {
   ModalBody,
   ModalCloseButton,
 } from "@chakra-ui/react";
-import { useState, useEffect, useCallback, useRef } from "react";
+import { useState, useCallback, useRef } from "react";
 import { Layout } from "~/components/layout";
 import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
@@ -54,63 +54,12 @@ import superjson from "superjson";
 import type { GetServerSideProps } from "next";
 import { useSession } from "next-auth/react";
 import { GPU_DISPLAY_NAMES } from "~/constants/gpu";
-import { generateStarterCode, DataTypes, Parameter } from "~/utils/starter";
+import { BenchmarkTestResult, SubmissionStatus, Submission } from "~/types/problem";
+import { useCodePersistence } from "~/components/problem/useCodePersistence";
+import { useSubmissionStream } from "~/hooks/useSubmissionStream";
+import { useSplitPanel } from "~/hooks/useSplitPanel";
+import { DataType, ProgrammingLanguage } from "~/types/misc";
 
-type BenchmarkTestResult = {
-  test_id: number;
-  runtime_ms: number;
-  gflops: number;
-  name: string;
-};
-
-type SubmissionStatus = {
-  status:
-    | "compiling"
-    | "CHECKING"
-    | "BENCHMARKING"
-    | "ACCEPTED"
-    | "WRONG_ANSWER"
-    | "ERROR"
-    | "SUBMISSIONS";
-  runtime: number | null;
-  gflops: number | null;
-  passedTests: number | null;
-  totalTests: number | null;
-  message: string | null;
-  errorMessage?: string;
-  errorDetails?: string;
-  benchmarkResults?: BenchmarkTestResult[];
-};
-
-type Submission = {
-  id: string;
-  status: string | null;
-  runtime: number | null;
-  gflops: number | null;
-  passedTests: number | null;
-  totalTests: number | null;
-  createdAt: Date;
-  problem: {
-    title: string;
-    slug: string;
-  };
-  gpuType: string;
-};
-
-const LOCAL_STORAGE_PREFIX = "problem_solution_";
-
-const getSolutionKey = (slug: string, language: string, dataType: string) => 
-  `${LOCAL_STORAGE_PREFIX}${slug}_${language}_${dataType}`;
-
-const saveSolutionToStorage = (slug: string, code: string, language: string, dataType: string) => {
-  if (typeof window === "undefined") return;
-  localStorage.setItem(getSolutionKey(slug, language, dataType), code);
-};
-
-const loadSolutionFromStorage = (slug: string, language: string, dataType: string): string | null => {
-  if (typeof window === "undefined") return null;
-  return localStorage.getItem(getSolutionKey(slug, language, dataType));
-};
 
 export const getServerSideProps: GetServerSideProps = async (context) => {
   const helpers = createServerSideHelpers({
@@ -145,55 +94,48 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
 };
 
 const getStatusMessage = (status: SubmissionStatus): string => {
-  if (status.message?.startsWith("status: CHECKING")) {
-    return "Checking...";
-  } else if (status.message?.startsWith("status: BENCHMARKING")) {
-    return "Running benchmarks...";
-  } else if (status.message?.startsWith("checker: compiling")) {
-    return "Compiling...";
-  } else if (status.message?.startsWith("checker: running") || status.message?.startsWith("checker: test_result")) {
-    return "Running tests...";
-  } else if (status.message?.startsWith("checker: complete")) {
-    return "Tests complete";
-  } else if (
-    status.message?.startsWith("benchmark: compiling") ||
-    status.message?.startsWith("benchmark: running") ||
-    status.message?.startsWith("benchmark: test_result")
-  ) {
-    return "Running benchmarks...";
-  } else if (status.message?.startsWith("benchmark: success")) {
-    return "Benchmark results";
-  } else if (status.message?.startsWith("complete: ACCEPTED")) {
-    return "ACCEPTED";
-  } else if (status.message?.startsWith("complete: WRONG_ANSWER")) {
-    return "Wrong answer";
-  } else if (status.message?.startsWith("error:")) {
-    return "Error";
+  if (!status.message) return `Status: ${status.status}`;
+  
+  if (status.message.startsWith("status: CHECKING")) return "Checking...";
+  if (status.message.startsWith("complete: ACCEPTED")) return "ACCEPTED";
+  
+  const messageMap: Record<string, string> = {
+    "status: BENCHMARKING": "Running benchmarks...",
+    "checker: compiling": "Compiling...",
+    "checker: running": "Running tests...",
+    "checker: test_result": "Running tests...",
+    "checker: complete": "Tests complete",
+    "benchmark: compiling": "Running benchmarks...",
+    "benchmark: running": "Running benchmarks...",
+    "benchmark: test_result": "Running benchmarks...",
+    "benchmark: success": "Benchmark results",
+    "complete: ACCEPTED": "ACCEPTED",
+    "complete: WRONG_ANSWER": "Wrong answer",
+    "error:": "Error",
+  };
+  
+
+  for (const [prefix, message] of Object.entries(messageMap)) {
+    if (status.message.startsWith(prefix)) return message;
   }
+  
   return `Status: ${status.status}`;
 };
 
 export default function ProblemPage({ slug }: { slug: string }) {
-  const { data: session, status } = useSession();
+    const { data: session } = useSession();
   const toast = useToast();
-  const [code, setCode] = useState("");
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [starterCode, setStarterCode] = useState("");
-  const [submissionStatus, setSubmissionStatus] =
-    useState<SubmissionStatus | null>(null);
-  const [submissionId, setSubmissionId] = useState<string | null>(null);
-  const [hasSetInitialCode, setHasSetInitialCode] = useState(false);
-  const [isTestCaseTableOpen, setIsTestCaseTableOpen] = useState(false);
-  const [isBenchmarking, setIsBenchmarking] = useState(false);
-  const [splitRatio, setSplitRatio] = useState(35);
-  const [isDragging, setIsDragging] = useState(false);
   const [selectedGpuType, setSelectedGpuType] = useState("T4");
-  const [isCodeDirty, setIsCodeDirty] = useState(false);
   const [isResetModalOpen, setIsResetModalOpen] = useState(false);
-  const [selectedLanguage, setSelectedLanguage] = useState("cuda");
-  const [selectedDataType, setSelectedDataType] = useState<DataTypes>("float32");
   const isProcessing = useRef<boolean>(false);
 
+  // Get problem data
+  const { data: problem, isLoading } = api.problems.getById.useQuery(
+    { slug },
+    { enabled: !!slug }
+  );
+
+  // Fetch submissions
   const submissionsQuery = api.problems.getSubmissions.useQuery(
     { problemSlug: slug },
     { enabled: !!slug }
@@ -203,13 +145,39 @@ export default function ProblemPage({ slug }: { slug: string }) {
     refetch: () => void;
   };
 
+  // Use custom hooks
+  const { splitRatio, isDragging, handleMouseDown, handleMouseUp } = useSplitPanel();
+  
+  const { 
+    code, 
+    setCode, 
+    selectedLanguage, 
+    setSelectedLanguage, 
+    selectedDataType, 
+    setSelectedDataType, 
+    isCodeDirty, 
+    handleReset, 
+    starterCode 
+  } = useCodePersistence(slug, problem);
+  
+  const {
+    isSubmitting,
+    submissionStatus,
+    submissionId,
+    isTestCaseTableOpen,
+    isBenchmarking,
+    setIsTestCaseTableOpen,
+    processSubmission,
+    startSubmission,
+    setSubmissionStatus
+  } = useSubmissionStream(submissionsQuery.refetch);
+
+  // Create submission mutation
   const createSubmissionMutation = api.problems.createSubmission.useMutation({
     onSuccess: (data) => {
-      setSubmissionId(data.id);
-      // No need to call submitMutation here - we'll directly connect to the streaming endpoint
+      void processSubmission(data.id);
     },
-    onError: (error) => {
-      setIsSubmitting(false);
+    onError: (error) => { 
       setSubmissionStatus(null);
       toast({
         title: "Failed to create submission",
@@ -221,7 +189,8 @@ export default function ProblemPage({ slug }: { slug: string }) {
     },
   });
 
-  const handleSubmit = () => {
+  // Handle submission
+  const handleSubmit = useCallback(() => {
     if (!session?.user) {
       toast({
         title: "Not signed in",
@@ -233,445 +202,24 @@ export default function ProblemPage({ slug }: { slug: string }) {
       return;
     }
 
-    setIsTestCaseTableOpen(false);
-    setIsBenchmarking(false);
-    setIsSubmitting(true);
-    setSubmissionStatus({
-      status: "CHECKING",
-      runtime: null,
-      gflops: null,
-      passedTests: null,
-      totalTests: null,
-      message: "status: CHECKING",
+    startSubmission();
+
+    createSubmissionMutation.mutate({
+      problemSlug: slug,
+      code: code,
+      language: selectedLanguage,
+      gpuType: selectedGpuType,
     });
-
-    createSubmissionMutation.mutate(
-      {
-        problemSlug: slug,
-        code,
-        language: "cuda",
-        gpuType: selectedGpuType,
-      },
-      {
-        onSuccess: (data) => {
-          // Wrap the async logic in a void function
-          void (async () => {
-            try {
-              const response = await fetch("/api/submissions/direct-submit", {
-                method: "POST",
-                headers: {
-                  "Content-Type": "application/json",
-                },
-                body: JSON.stringify({ submissionId: data.id }),
-                cache: "no-store",
-                credentials: "same-origin",
-                keepalive: true,
-                signal: AbortSignal.timeout(300000),
-              });
-
-              if (!response.ok) {
-                throw new Error(
-                  `Direct submit API returned ${response.status}`
-                );
-              }
-
-              // Set up reader for streaming response
-              const reader = response.body?.getReader();
-              if (!reader) {
-                throw new Error("No response body from direct-submit");
-              }
-
-              // Process the stream
-              const decoder = new TextDecoder();
-              let buffer = "";
-
-              const MAX_RETRIES = 3;
-              let retryCount = 0;
-
-              const attemptConnection = async (): Promise<void> => {
-                try {
-                  while (true) {
-                    const { done, value } = await reader.read();
-                    if (done) {
-                      console.log("[sse] Stream complete");
-                      break;
-                    }
-
-                    // Reset retry count on successful read
-                    retryCount = 0;
-
-                    const chunk = decoder.decode(value, { stream: true });
-                    buffer += chunk;
-
-                    // Process complete events in buffer
-                    const events = buffer.split("\n\n");
-                    buffer = events.pop() ?? ""; // Use nullish coalescing
-
-                    for (const event of events) {
-                      if (!event) continue;
-
-                      const eventLines = event.split("\n");
-                      let eventType = "message";
-                      let eventData = "";
-
-                      for (const line of eventLines) {
-                        if (line.startsWith("event: ")) {
-                          eventType = line.slice(7);
-                        } else if (line.startsWith("data: ")) {
-                          eventData = line.slice(6);
-                        }
-                      }
-
-                      if (!eventData) continue;
-
-                      try {
-                        // Parse and type the data
-                        const data = JSON.parse(eventData) as {
-                          status?: string;
-                          passedTests?: number;
-                          totalTests?: number;
-                          result?: {
-                            status?: string;
-                            test_id?: number;
-                            runtime_ms?: number;
-                            gflops?: number;
-                            name?: string;
-                          };
-                          runtime?: number;
-                          gflops?: number;
-                          error?: string;
-                          details?: string;
-                          benchmarkResults?: BenchmarkTestResult[];
-                        };
-
-                        console.log(`[sse] ${eventType} event:`, data);
-
-                        // Handle different event types
-                        if (eventType === "status") {
-                          setSubmissionStatus((prev) => {
-                            if (!prev) return prev;
-                            return {
-                              ...prev,
-                              status:
-                                (data.status as SubmissionStatus["status"]) ??
-                                prev.status,
-                              passedTests: data.passedTests ?? prev.passedTests,
-                              totalTests: data.totalTests ?? prev.totalTests,
-                              message:
-                                eventType +
-                                ": " +
-                                (data.status ?? prev.message),
-                            };
-                          });
-                        } else if (eventType === "checker") {
-                          if (data.status === "test_result" && data.result) {
-                            setSubmissionStatus((prev) => {
-                              if (!prev) {
-                                return {
-                                  status: "CHECKING",
-                                  runtime: null,
-                                  gflops: null,
-                                  passedTests:
-                                    data.result?.status === "PASSED" ? 1 : 0,
-                                  totalTests: data.totalTests ?? 1,
-                                  message: eventType + ": " + data.status,
-                                };
-                              }
-                              return {
-                                ...prev,
-                                passedTests:
-                                  (prev.passedTests ?? 0) +
-                                  (data.result?.status === "PASSED" ? 1 : 0),
-                                totalTests: data.totalTests ?? prev.totalTests,
-                                message: eventType + ": " + data.status,
-                              };
-                            });
-                          } else if (data.status && data.status !== "error") {
-                            // Handle other checker statuses like "compiling"
-                            setSubmissionStatus((prev) => {
-                              if (!prev) return prev;
-                              return {
-                                ...prev,
-                                message: eventType + ": " + data.status,
-                              };
-                            });
-                          } else if (data.status === "error") {
-                            setSubmissionStatus({
-                              status: "ERROR",
-                              runtime: null,
-                              gflops: null,
-                              passedTests: null,
-                              totalTests: null,
-                              message:
-                                eventType + ": " + (data.status ?? "ERROR"),
-                              errorMessage: data.error ?? undefined,
-                              errorDetails: data.details ?? undefined,
-                            });
-
-                            setIsSubmitting(false);
-                            submissionsQuery.refetch();
-
-                            toast({
-                              title: "Submission Error",
-                              description:
-                                data.error ??
-                                "An error occurred during submission",
-                              status: "error",
-                              duration: 5000,
-                              isClosable: true,
-                            });
-                            return;
-                          }
-                        } else if (eventType === "benchmark") {
-                          if (data.status === "test_result" && data.result) {
-                            setIsBenchmarking(true);
-                            setIsTestCaseTableOpen(true);
-                            setSubmissionStatus((prev) => {
-                              if (!prev) return prev;
-                              const benchmarkResult: BenchmarkTestResult = {
-                                test_id: data.result?.test_id ?? 0,
-                                runtime_ms: data.result?.runtime_ms ?? 0,
-                                gflops: data.result?.gflops ?? 0,
-                                name:
-                                  data.result?.name ??
-                                  `Test ${data.result?.test_id ?? 0}`,
-                              };
-                              return {
-                                ...prev,
-                                message: eventType + ": " + data.status,
-                                benchmarkResults: [
-                                  ...(prev.benchmarkResults ?? []),
-                                  benchmarkResult,
-                                ],
-                              };
-                            });
-                          } else if (data.status && data.status !== "error") {
-                            // Handle other benchmark statuses
-                            setSubmissionStatus((prev) => {
-                              if (!prev) return prev;
-                              return {
-                                ...prev,
-                                message: eventType + ": " + data.status,
-                              };
-                            });
-                          }
-                        } else if (eventType === "complete") {
-                          setSubmissionStatus({
-                            status:
-                              (data.status as SubmissionStatus["status"]) ??
-                              "ERROR",
-                            runtime: data.runtime ?? null,
-                            gflops: data.gflops ?? null,
-                            passedTests: data.passedTests ?? null,
-                            totalTests: data.totalTests ?? null,
-                            message:
-                              eventType + ": " + (data.status ?? "ERROR"),
-                            errorMessage: data.error ?? undefined,
-                            errorDetails: data.details ?? undefined,
-                            benchmarkResults:
-                              data.benchmarkResults ?? undefined,
-                          });
-
-                          setIsSubmitting(false);
-                          submissionsQuery.refetch();
-                          return;
-                        } else if (eventType === "error") {
-                          setSubmissionStatus({
-                            status: "ERROR",
-                            runtime: null,
-                            gflops: null,
-                            passedTests: null,
-                            totalTests: null,
-                            message:
-                              eventType + ": " + (data.status ?? "ERROR"),
-                            errorMessage: data.error ?? undefined,
-                            errorDetails: data.details ?? undefined,
-                          });
-
-                          setIsSubmitting(false);
-                          submissionsQuery.refetch();
-
-                          toast({
-                            title: "Submission Error",
-                            description:
-                              data.error ??
-                              "An error occurred during submission",
-                            status: "error",
-                            duration: 5000,
-                            isClosable: true,
-                          });
-                          return;
-                        }
-                      } catch (error) {
-                        console.error(
-                          "Error parsing event data:",
-                          error,
-                          "Raw data:",
-                          eventData
-                        );
-                      }
-                    }
-                  }
-
-                  // If we reach here, the stream ended normally
-                  setIsSubmitting(false);
-                } catch (error) {
-                  console.error("[sse] Stream error:", error);
-
-                  // Attempt retry if we haven't exceeded max retries
-                  if (retryCount < MAX_RETRIES) {
-                    retryCount++;
-                    console.log(
-                      `[sse] Retrying connection (${retryCount}/${MAX_RETRIES})...`
-                    );
-
-                    // Wait before retrying (exponential backoff)
-                    await new Promise((resolve) =>
-                      setTimeout(resolve, 1000 * Math.pow(2, retryCount))
-                    );
-
-                    return attemptConnection();
-                  } else {
-                    throw error; // Re-throw if max retries exceeded
-                  }
-                }
-              };
-
-              void attemptConnection();
-            } catch (error) {
-              console.error("[sse] Error:", error);
-              setIsSubmitting(false);
-
-              let errorMessage = "Failed to connect to submission service";
-              if (error instanceof Error) {
-                if (
-                  error.message.includes("QUIC_PROTOCOL_ERROR") ||
-                  error.message.includes("network error") ||
-                  error.message.includes("failed to fetch")
-                ) {
-                  errorMessage =
-                    "Network protocol error. This may be due to your connection or a server configuration issue. Please try again in a few moments.";
-                } else {
-                  errorMessage = error.message;
-                }
-              }
-
-              toast({
-                title: "Connection Error",
-                description: errorMessage,
-                status: "error",
-                duration: 5000,
-                isClosable: true,
-              });
-            }
-          })();
-        },
-      }
-    );
-  };
-
-  const { data: problem, isLoading } = api.problems.getById.useQuery(
-    { slug: slug },
-    { enabled: !!slug }
-  );
-
-  const getStarterCode = useCallback(() => {
-    return generateStarterCode(problem?.parameters as unknown as Parameter[], selectedLanguage, selectedDataType);
-  }, [problem?.parameters, selectedLanguage, selectedDataType]);
-
-
-  useEffect(() => {
-    if (!hasSetInitialCode && slug) {
-      const savedSolution = loadSolutionFromStorage(slug, selectedLanguage, selectedDataType);
-      if (savedSolution) {
-        setCode(savedSolution);
-        setHasSetInitialCode(true);
-      } else if (starterCode) {
-        setCode(starterCode);
-        setHasSetInitialCode(true);
-      }
-    }
-  }, [slug, hasSetInitialCode, problem, selectedLanguage, selectedDataType]);
-
-  useEffect(() => {
-    if (code && slug) {
-      saveSolutionToStorage(slug, code, selectedLanguage, selectedDataType);
-    }
-  }, [code, slug]);
-
-  useEffect(() => {
-    if (starterCode) {
-      setIsCodeDirty(code !== starterCode);
-    }
-  }, [code, starterCode]);
-
-  useEffect(() => {
-    const newStarterCode = generateStarterCode(problem?.parameters as unknown as Parameter[], selectedLanguage, selectedDataType);
-    if (newStarterCode) {
-      setStarterCode(newStarterCode);
-      
-      const savedSolution = loadSolutionFromStorage(slug, selectedLanguage, selectedDataType);
-      if (savedSolution) {
-        setCode(savedSolution);
-      } else {
-        setCode(newStarterCode);
-        if (slug) {
-          saveSolutionToStorage(slug, newStarterCode, selectedLanguage, selectedDataType);
-        }
-      }
-    }
-  }, [selectedLanguage, selectedDataType, problem?.parameters, slug]);
-
-  const handleReset = () => {
-    if (starterCode) {
-      setCode(starterCode);
-      // Clear localStorage only for current language/datatype combination
-      if (slug) {
-        localStorage.removeItem(getSolutionKey(slug, selectedLanguage, selectedDataType));
-      }
-    }
-  };
-
-  // Move these handlers to useCallback
-  const handleMouseDown = useCallback(() => {
-    setIsDragging(true);
-    document.body.style.cursor = "col-resize";
-    document.body.style.userSelect = "none";
-  }, []);
-
-  const handleMouseUp = useCallback(() => {
-    setIsDragging(false);
-    document.body.style.cursor = "default";
-    document.body.style.userSelect = "auto";
-  }, []);
-
-  useEffect(() => {
-    if (isDragging) {
-      // Move handleMouseMove inside useEffect
-      const handleMouseMove = (e: MouseEvent) => {
-        if (!isDragging) return;
-
-        const container = document.getElementById("split-container");
-        if (!container) return;
-
-        const containerRect = container.getBoundingClientRect();
-        const newRatio =
-          ((e.clientX - containerRect.left) / containerRect.width) * 100;
-
-        // Adjust max ratio to 55% to ensure submit button stays visible
-        const clampedRatio = Math.min(Math.max(newRatio, 35), 55);
-        setSplitRatio(clampedRatio);
-      };
-
-      window.addEventListener("mousemove", handleMouseMove);
-      window.addEventListener("mouseup", handleMouseUp);
-
-      return () => {
-        window.removeEventListener("mousemove", handleMouseMove);
-        window.removeEventListener("mouseup", handleMouseUp);
-      };
-    }
-  }, [isDragging, handleMouseUp]);
+  }, [
+    session?.user, 
+    slug, 
+    code, 
+    selectedLanguage, 
+    selectedGpuType, 
+    createSubmissionMutation, 
+    startSubmission, 
+    toast
+  ]);
 
   if (isLoading) {
     return (
@@ -697,6 +245,7 @@ export default function ProblemPage({ slug }: { slug: string }) {
       </Layout>
     );
   }
+
 
   return (
     <Layout title={problem.title}>
@@ -1041,7 +590,7 @@ export default function ProblemPage({ slug }: { slug: string }) {
                                 </Thead>
                                 <Tbody>
                                   {submissionStatus.benchmarkResults?.map(
-                                    (result) => (
+                                    (result: BenchmarkTestResult) => (
                                       <Tr
                                         key={result.test_id}
                                         _hover={{ bg: "whiteAlpha.100" }}
@@ -1396,10 +945,9 @@ export default function ProblemPage({ slug }: { slug: string }) {
                     bg="whiteAlpha.50"
                     borderColor="whiteAlpha.200"
                     _hover={{ borderColor: "whiteAlpha.300" }}
-                    onChange={(e) => setSelectedLanguage(e.target.value)}
+                    onChange={(e) => setSelectedLanguage(e.target.value as ProgrammingLanguage)}
                     value={selectedLanguage}
                     w="160px"
-                    defaultValue="cuda"
                     borderRadius="full"
                     sx={{
                       "& > option": {
@@ -1424,7 +972,7 @@ export default function ProblemPage({ slug }: { slug: string }) {
                     _hover={{ borderColor: "whiteAlpha.300" }}
                     w="140px"
                     value={selectedDataType}
-                    onChange={(e) => setSelectedDataType(e.target.value as DataTypes)}
+                    onChange={(e) => setSelectedDataType(e.target.value as DataType)}
                     borderRadius="full"
                     sx={{
                       "& > option": {
