+++ src/types/misc.ts
@@ -1,0 +1,3 @@
+export type ProgrammingLanguage = "cpp" | "cuda" | "python"
+
+export type DataType = "float16" | "float32" | "int32" | "int16"
